+++ service/persist/feature.go
@@ -9,7 +9,7 @@ import (
 
 // FeatureFlag represents a feature flag in the database
 type FeatureFlag struct {
-	Version      NullInt64       `json:"version"` // schema version for this model
+	Version      NullInt32       `json:"version"` // schema version for this model
 	ID           DBID            `json:"id" binding:"required"`
 	CreationTime CreationTime    `json:"created_at"`
 	Deleted      NullBool        `json:"-"`
