+++ service/persist/nft.go
@@ -9,7 +9,7 @@ import (
 
 // NFT represents an nft throughout the application
 type NFT struct {
-	Version         NullInt64       `json:"version"` // schema version for this model
+	Version         NullInt32       `json:"version"` // schema version for this model
 	ID              DBID            `json:"id" binding:"required"`
 	CreationTime    CreationTime    `json:"created_at"`
 	Deleted         NullBool        `json:"-"`
